+++ Libraries/PeriphDrivers/Source/TMR/tmr_me21.c
@@ -26,7 +26,7 @@
 int MXC_TMR_Init(mxc_tmr_regs_t *tmr, mxc_tmr_cfg_t *cfg, bool init_pins)
 {
     uint8_t tmr_id = MXC_TMR_GET_IDX(tmr);
-    uint8_t clockSource = MXC_TMR_CLK0;
+    int clockSource = MXC_TMR_CLK0;
 
     if (cfg == NULL) {
         return E_NULL_PTR;
